clear up false positive comment and clean up code
This is not a buffer underflow because fgets the line can never be empty
if we have a line, we want to terminate the line on the new line break.
This code makes it much clearer what we are doing and why.
In the event that we have no new line, we also will no longer truncate the
last byte of our line.

We will exit the loop if we do not meet these conditions.